requirements.txt
flask
flask-cors
pillow
numpy
tensorflow
keras
